endpoint 수정
